Add userInput "rock" conditionals
Players win, lose, or draw. When a player wins a
victory message is displayed in console and their
score increases by one. When they lose a loss
message appears in console and the computer's scores
increases by one. In a draw a message is shown
in console. Current score tallies are printed
to the console as well.